chore: move featureFlags into constants
